+++ frontend/package.json
@@ -14,10 +14,14 @@
     "@emotion/styled": "^11.10.6",
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.12.1",
+    "localforage": "^1.10.0",
+    "match-sorter": "^6.3.1",
     "prompt-sync": "^4.2.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "reactflow": "^11.7.0"
+    "react-router-dom": "^6.11.1",
+    "reactflow": "^11.7.0",
+    "sort-by": "^1.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
